chore: format and clippy suggestions

+++ src/image_processing.rs
@@ -113,7 +113,7 @@ impl Position {
     pub fn normalized_len(&self) -> f32 {
         let move_speed = CONFIG.neural_network.retina_movement_speed as i32;
         let max_len = ((move_speed.pow(2) + move_speed.pow(2)) as f32).sqrt();
-        self.len() as f32 / max_len
+        self.len() / max_len
     }
 
     pub fn x(&self) -> i32 {
